+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.4
+`2017年12月21日`
+
+1. 更新tinyjs-plugin-mesh 为0.0.4 Mesh 新增了 canvasDrawTimes属性
+
+2. 给TinyArmatureDisplay增加 meshCanvasRenderDrawImageTimes属性
+
+3. 解决mesh canvas render 渲染骨骼网格动画的时候空白间隙 gap line的问题
 
 ## 0.1.3
 `2017-12-20`